infinity scroll applied

+++ face/app/app.py
@@ -21,31 +21,38 @@ def render_file():
 def upload_file():
     if request.method == 'POST':
         f = request.files['file']
+
+        ext = secure_filename(f.filename).split('.')[-1]
+        if ext != 'jpg' and ext != 'png' :
+            return 'please post jpg or png image', 400
+        
         image_ori = os.path.join(PATH_TEMP, secure_filename(f.filename))
-        image_thm = os.path.join(PATH_TEMP, 'thm_'+ secure_filename(f.filename).split('.')[-2] + '.jpg')
+        # image_thm = os.path.join(PATH_TEMP, 'thm_'+ secure_filename(f.filename).split('.')[-2] + '.jpg')
     
         if not os.path.exists(PATH_TEMP):
             os.makedirs(PATH_TEMP)
 
         f.save(image_ori)
-        img = Image.open(image_ori)
-        img.thumbnail(THM_SIZE)
-        img.convert('RGB').save(image_thm)
+        # img = Image.open(image_ori)
+        # img.thumbnail(THM_SIZE)
+        # img.convert('RGB').save(image_thm)
                     
-        who = get_face_name(image_thm)
+        who = get_face_name(image_ori)
 
         try:
             os.remove(image_ori)
         except OSError:
             pass
 
-        try:
-            os.remove(image_thm)
-        except OSError:
-            pass
+        # try:
+        #     os.remove(image_thm)
+        # except OSError:
+        #     pass
 
         return jsonify({'who':list(set(who))})
+    return 'please post jpg or png image', 400
 
 
 if __name__ == '__main__':
-    app.run(host='0.0.0.0', debug=True, port=4000)
+    port = os.environ.get('FACE_PORT')
+    app.run(host='0.0.0.0', debug=True, port=port)
